refactored the client support into seperate crates that implement the cache inside a wrapper
