+++ .gitignore
@@ -1,5 +1,5 @@
 /target
 Cargo.lock
 **/*.rs.bk
-/http-cacache
+http-cacache/
 /.idea
