+++ http-cache-reqwest/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "http-cache-types"
+name = "http-cache-reqwest"
 version = "0.1.0"
-description = "Common types used by http-cache"
+description = "http-cache middleware implementation for reqwest"
 authors = ["Christian Haynes <06chaynes@gmail.com>", "Kat Marchán <kzm@zkat.tech>"]
 repository = "https://github.com/06chaynes/http-cache.git"
 license = "MIT OR Apache-2.0"
@@ -16,20 +16,26 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0"
 async-trait = "0.1"
-bincode = { version = "1.3", optional = true }
-cacache = { version = "9.0", optional = true }
 http = "0.2"
 http-cache-semantics = "0.9"
-httpdate = "1.0"
-http-types = { version = "2.12", optional = true }
-miette = "3.3"
-reqwest = { version = "0.11", default-features = false, optional = true }
-reqwest-middleware = { version = "0.1", optional = true }
+reqwest = { version = "0.11", default-features = false}
+reqwest-middleware = "0.1"
 serde = { version = "1.0", features = ["derive"] }
-thiserror = "1.0"
+task-local-extensions = "0.1"
 url = { version = "2.2", features = ["serde"] }
 
+[dependencies.http-cache]
+path = "../."
+
+[dev-dependencies]
+cacache = "9.0"
+tokio = { version = "1.15", features = ["macros", "rt-multi-thread"] }
+mockito = "0.30"
+
+[dev-dependencies.http-cache]
+path = "../."
+features = ["manager-cacache"]
+
 [features]
-manager-cacache = ["cacache", "bincode"]
-client-reqwest = ["reqwest", "reqwest-middleware"]
-client-surf = ["http-types"]
+default = ["manager-cacache"]
+manager-cacache = ["http-cache/manager-cacache"]
