+++ http-cache-surf/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "http-cache-middleware-surf"
+name = "http-cache-surf"
 version = "0.1.0"
 description = "http-cache middleware implementation for surf"
 authors = ["Christian Haynes <06chaynes@gmail.com>", "Kat Marchán <kzm@zkat.tech>"]
@@ -23,6 +23,19 @@ serde = { version = "1.0", features = ["derive"] }
 surf = "2.3"
 url = { version = "2.2", features = ["serde"] }
 
-[dependencies.http-cache-types]
-path = "../http-cache-types"
-features = ["client-surf"]
+[dependencies.http-cache]
+path = "../."
+features = ["with-http-types"]
+
+[dev-dependencies]
+async-std = { version = "1.10", features = ["attributes"]}
+cacache = "9.0"
+mockito = "0.30"
+
+[dev-dependencies.http-cache]
+path = "../."
+features = ["manager-cacache"]
+
+[features]
+default = ["manager-cacache"]
+manager-cacache = ["http-cache/manager-cacache"]
