+++ src/managers/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "manager-cacache")]
+pub(crate) mod cacache;
